+++ structures/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = 'structures.apps.StructuresConfig'
 
-__version__ = '0.2.0'
+__version__ = '0.2.1'
 __title__ = 'Alliance Structures'